Move pkg/library/storage to pkg/provision/storage
