+++ pkg/provision/storage/testdata/rewrites-volumes-for-common-pvc-strategy.yaml
